requirements.txt
bottle>=0.12.8
bottle-sqlite>=0.1.2
requests>=2.6.0
gunicorn>=19.3.0
pynacl